Refactor DataTable (VueTable-2)
