+++ src/views/shipment/index.vue
@@ -19,7 +19,7 @@
         </nav>
         <vuetable ref="vuetable"
           api-url="/api/v2/parcels/web"
-          :css="css"
+          :css="bulmaTableCss"
           :fields="table.columns"
           :paginationPath="paginationPath"
           :data-path="dataPath"
@@ -41,20 +41,20 @@
 
 <script>
 import Vue from 'vue';
-import DetailRow from './DetailRow';
 
-import Status from './Status';
-import tableMixin from '../../components/Table/mixin';
+import dataTableMixin from '../../mixins/dataTable';
 import Vuetable from 'vuetable-2/src/components/Vuetable.vue';
 import VuetablePaginationInfo from 'vuetable-2/src/components/VuetablePaginationInfo';
-import FilterBar from '../../components/Table/FilterBar.vue';
-import BulmaPagination from '../../components/Table/BulmaPagination.vue';
+import FilterBar from '../../components/FilterBar.vue';
+import BulmaPagination from '../../components/BulmaPagination.vue';
+import DetailRow from './DetailRow';
+import Status from './Status';
 
 Vue.component('shipment-detail-row', DetailRow)
 Vue.component('shipment-status', Status)
 
 export default {
-  mixins: [tableMixin],
+  mixins: [dataTableMixin],
   components: {
     Vuetable,
     BulmaPagination,
